Made all talons use the same forward direction.
Essentially, a static final boolean field is used to set all the drive
talons as inverted or not. The key point though is that they are either
all inverted or all normal. So, they all will have the same direction as
forward, and the same direction for backward.

Also, I checked, and the Differential Drive class internally handles
reversing the direction of the right side wheels. So when it's told to
go forward it actually goes forward and doesn't just spin in a circle.
My synchronizing of the forward/reverse directions should not affect it.

Signed-off-by: Loren Frates <cubedbits@outlook.com>