Refactored controls to allow switching
Essentially, on the dashboard the driver can select how the robot will operate and the robot will switch to that control scheme. They can change it as many times as they like. Currently there are two control schemes available: PositionBasedControl, and ThrottleBasedControl. PositionBasedControl allows the drivers to change the position setpoints of most components. The more basic ones are controlled only by throttle (like the drive). ThrottleBasedControl allows all motors to be controlled only be throttle. This mode is pretty much a global motor-box.